better seed of orders, fixes for alerts and groups on backend, advanced column manager in datatable, variable for buttons in datatable, fix for 0 in form resolver, more components in form, and slot for custom components, value modifiers in ObjectView, fixes for lists: alertTypes, groups, users and new lists for alerts, displays and orders

+++ BusinessLogic/routes/Groups.js
@@ -141,6 +141,7 @@ router.put('/:id', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
     if (!req.decodedToken.admin) return res.status(403).json({ error: 'Unauthorized: Admin privileges required' });
+    if (req.params.id == 1) return res.status(403).json({ error: 'Unauthorized: Cannot delete main group' });
 
     try {
         await Groups.destroy({
